requirements.txt
# Fire & Ice model — data and backtest
pandas>=1.5.0
numpy>=1.20.0
pyyaml>=6.0
requests>=2.28.0
yfinance>=0.2.28
pyarrow>=10.0.0
